+++ flowcelltool/flowcells/views.py
@@ -322,7 +322,7 @@ class FlowCellListView(
 
     #: Flow cells are sorted by run date (inferred from the name when being
     #: created, latest come first)
-    queryset = models.FlowCell.objects.order_by('-run_date')
+    queryset = models.FlowCell.objects.order_by('-run_date', '-run_number', '-slot')
 
     #: Pagination with 50 items should work fine for us
     paginate_by = 50