+++ HISTORY.rst
@@ -6,6 +6,8 @@ History
 HEAD
 ----
 
+- Fixing flow cell order.
+- Fixing bug in flow cell list.
 - Changing API ``flowcell/by_vendor_id`` to ``flowcell_resolve``.
 
 ------